fix/mitigation:  Disable `keyDragMoveStacks` -- Unsure why but even having this hotkey bound and pressed causes a JVM Lock / Crash.

+++ src/main/java/fi/dy/masa/itemscroller/util/AccessorUtils.java
@@ -18,7 +18,7 @@ public class AccessorUtils
     }
 
     @Nullable
-    public static Slot getSlotAtPosition(HandledScreen<?> gui, int x, int y)
+    public static Slot getSlotAtPosition(HandledScreen<?> gui, double x, double y)
     {
         return ((IMixinScreenWithHandler) gui).itemscroller_getSlotAtPositionInvoker(x, y);
     }
